.env file added with environment variables for database
