+++ .env
@@ -1,0 +1,5 @@
+DATABASE_HOSTNAME = localhost
+DATABASE_PORT = 5432
+DATABASE_PASSWORD = Prolific1
+DATABASE_NAME = level_up
+DATABASE_USERNAME = postgres
